components: add HOC component to read query state
